完善 breeze:publish 增加 --gql 选项只更新gql

+++ src/Console/PublishCommand.php
@@ -14,7 +14,7 @@ class PublishCommand extends Command
      *
      * @var string
      */
-    protected $signature = 'breeze:publish {--force}';
+    protected $signature = 'breeze:publish {--force} {--gql}';
 
     /**
      * The Console command description.
@@ -32,13 +32,28 @@ class PublishCommand extends Command
     {
         $force = $this->option('force');
 
+        if ($gql = $this->option('gql')) {
+            $this->info("只发布gql...");
+
+            // 清理breeze分散的gqls
+            if (is_dir(base_path('graphql/article'))) {
+                File::deleteDirectory(base_path('graphql'));
+            }
+
+            $this->call('vendor:publish', ['--tag' => 'media-graphql', '--force' => $force]);
+            $this->call('vendor:publish', ['--tag' => 'content-graphql', '--force' => $force]);
+            $this->call('vendor:publish', ['--tag' => 'sns-graphql', '--force' => $force]);
+            $this->call('vendor:publish', ['--tag' => 'task-graphql', '--force' => $force]);
+            $this->call('vendor:publish', ['--tag' => 'wallet-graphql', '--force' => $force]);
+            $this->call('vendor:publish', ['--tag' => 'question-graphql', '--force' => $force]);
+            $this->call('vendor:publish', ['--tag' => 'breeze-graphql', '--force' => $force]);
+            return;
+        }
+
         // lighthouse 和 playground
         // gqlp配置 后面自己手动修改
         $this->callSilent("vendor:publish", ["--provider" => "Nuwave\Lighthouse\LighthouseServiceProvider", '--force' => $force]);
         $this->callSilent("vendor:publish", ["--provider" => "MLL\GraphQLPlayground\GraphQLPlaygroundServiceProvider", '--force' => $force]);
-
-        // 清理breeze分散的gqls
-        File::deleteDirectory(base_path('graphql'));
 
         // breeze子模块的资源配置
         $this->callSilent('media:publish', ['--force' => $force]);
